Rename div to 'call-action' and add data to it
As there were buttons with the value 'sign-up' on the index page, it was
bound to cause confusion in the future when making changes to a class
that also had also had the name 'sign-up'.

The class name of the div has been changed to 'call-action' to make it
clear that we are referring to one of the div's on the webpage.