Refactor information section to separate card divs
All the card divs in the information section are currently not in their
own div, which makes it difficult to style other specific parts of
the section.

Refactored the information section to include a new div that contains
all of the card divs, making it easier to style the section.